requirements.txt
click==8.1.3
flask==2.2.3
greenlet==2.0.2
itsdangerous==2.1.2
jinja2==3.1.2
markupsafe==2.1.2
numpy==1.24.2
pandas==1.5.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.2
six==1.16.0
sqlalchemy==2.0.7
typing-extensions==4.5.0
werkzeug==2.2.3